Add files via upload
These are all the files with everyones functions to make a 'working' text based game. I added breaks for the while loop in mountainDay and convocation_debugging so they wouldn't keep repeating. Added score to file. Need to add appending to $score to mountainDay and convocation so it can be accounted for when picking major.