Add files via upload
Added images so when user makes a choice, it would move on. I also added the scores so it would give the user a score at the end but the number still needs to be changed. There's an issue of when the user clicks outside of an image it continues or circles back. I tried fixing it by calling the class again or changing $state to a high number so clicking would do nothing, but it doesn't seem to work.