Add files via upload
The addition of the mountain day files that have been edited. I moved process_click() so it would be in the classes since before they were just function right outside of their respective classes. i also moved the $state of some of the variables for the mouse_down portion since a few were incorrect and it led to errors.