Addition of downtown
I added downtown to the file and to do that I had to remove the mouse_down() for the downtown_main() function to mountain-main by changing the $state  number. When I went to Waka's TA hours she mentioned that me may only be able to use one mouse_down() so I switched it so it was in one file.