Adding downtown
It seems like everything works all together and that scores is getting everything from all the files. I'll need to change the average to be in the middle again and will do that later. I also need to change the formatting of everything so its consistent and change any typos.